test mac build (6)

+++ project/src/libwbfs/libwbfs.c
@@ -398,7 +398,7 @@ int wbfs_calc_size_shift
 	)
     {
 	// ensure that wbfs_sec_sz is big enough to address every blocks using 16 bits
-	if ( n_wii_sec < (u32)WBFS_MAX_SECTORS << shift_count )
+	if ( n_wii_sec < WBFS_MAX_SECTORS << shift_count )
 	    break;
     }
